flags work as far as i can tell
