--- server/api/user/user.socket.js
@@ -1,27 +1,0 @@
-'use strict';
-
-var User = require('./user.model');
-
-exports.register = function(socket) {
-  User.schema.post('save', function (doc) {
-    onSave(socket, doc);
-  });
-  User.schema.post('remove', function (doc) {
-    onRemove(socket, doc);
-  });
-  User.schema.put('save', function (doc) {
-    onUpdate(socket, doc);
-  });
-}
-
-function onSave(socket, doc, cb) {
-  socket.emit('user:save', doc);
-}
-
-function onRemove(socket, doc, cb) {
-  socket.emit('user:remove', doc);
-}
-
-function onUpdate(socket, doc, cb) {
-  socket.emit('user:'+doc._id, doc);
-}
